requirements.txt
requests
telebot
python-dotenv